requirements.txt
flask==3.0.0
google-genai==0.2.2
Pillow==10.1.0
python-dotenv==1.0.0
pytesseract
google-genai
dotenv
SpeechRecognition
pyaudio